Enable the V8 Virtual Memory Cage by Default
The goal of the V8 heap sandbox is to prevent an attacker with the
ability to corrupt V8 objects from constructing an arbitrary memory
read/write primitive. This is achieved through two fundamental
mechanisms: the virtual memory cage, discussed in this document,
and the external pointer table.

Upstream is in the process of testing enabling this.

See:
* https://twitter.com/5aelo/status/1450862485070827533
* https://chromium-review.googlesource.com/c/chromium/src/+/3220290